+++ cypress/integration/route/PlayOffline.spec.ts
@@ -17,9 +17,10 @@ describe("/play-offline", () => {
     Fari.get(
       "scene.player-row.gm-npc-0.assign-or-open-character-sheet"
     ).click();
-    Fari.get("character-dialog.name").type("Luke Skywalker");
+    Fari.get("character-dialog.name").clear().type("Luke Skywalker");
     Fari.waitContentEditable();
     Fari.get("character-dialog.save").click();
+    Fari.get("character-dialog.close").click();
 
     // add leia
     Fari.get("scene.add-player").click();
@@ -33,9 +34,10 @@ describe("/play-offline", () => {
     Fari.get(
       "scene.player-row.gm-npc-1.assign-or-open-character-sheet"
     ).click();
-    Fari.get("character-dialog.name").type("Leia Organa");
+    Fari.get("character-dialog.name").clear().type("Leia Organa");
     Fari.waitContentEditable();
     Fari.get("character-dialog.save").click();
+    Fari.get("character-dialog.close").click();
 
     // play
     Fari.get("scene.player-row.gm-npc-0.toggle-initiative").click();