+++ lib/routes/Bugs/BugsRoute.tsx
@@ -1,7 +1,7 @@
 import Box from "@material-ui/core/Box";
 import Container from "@material-ui/core/Container";
 import { useTheme } from "@material-ui/core/styles";
-import QuestionAnswerIcon from "@material-ui/icons/QuestionAnswer";
+import BugReportIcon from "@material-ui/icons/BugReport";
 import React from "react";
 import { Heading } from "../../components/Heading/Heading";
 import { Page } from "../../components/Page/Page";
@@ -9,28 +9,28 @@ import { PageMeta } from "../../components/PageMeta/PageMeta";
 import { useCanny } from "../../hooks/useCanny/useCanny";
 import { useTranslate } from "../../hooks/useTranslate/useTranslate";
 
-export function FeatureRequestsRoute() {
+export function BugsRoute() {
   const { t } = useTranslate();
   const theme = useTheme();
   const canny = useCanny({
-    boardToken: "7af01bfe-e4a0-f03f-7eae-48a7edc219da",
-    basePath: "/feature-requests",
+    boardToken: "0e1da373-6ec5-a2b8-ca4f-893c4d415990",
+    basePath: "/bugs",
   });
 
   const pageTitle = canny.postTitle
-    ? `${canny.postTitle} | ${t("feature-requests-route.meta.title")}`
-    : t("feature-requests-route.meta.title");
+    ? `${canny.postTitle} | ${t("bugs-route.meta.title")}`
+    : t("bugs-route.meta.title");
 
   return (
     <Page>
       <PageMeta
         title={pageTitle}
-        description={t("feature-requests-route.meta.description")}
+        description={t("bugs-route.meta.description")}
       />
       <Heading
-        icon={QuestionAnswerIcon}
-        title={t("feature-requests-route.meta.title")}
-        subtitle={t("feature-requests-route.meta.description")}
+        icon={BugReportIcon}
+        title={t("bugs-route.meta.title")}
+        subtitle={t("bugs-route.meta.description")}
       />
       <Container maxWidth="lg">
         <Box bgcolor="#fff" px="2rem" py="2rem" boxShadow={theme.shadows[1]}>
@@ -41,4 +41,4 @@ export function FeatureRequestsRoute() {
   );
 }
 
-export default FeatureRequestsRoute;
+export default BugsRoute;