+++ stories/CharacterSheet.stories.tsx
@@ -82,203 +82,91 @@ const Template: Story<IProps> = (args, context) => {
   );
 };
 
-export const FateCondensed = Template.bind({});
-(FateCondensed as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.FateCondensed
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const FateCore = Template.bind({});
-(FateCore as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(CharacterTemplates.FateCore);
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const FateAccelerated = Template.bind({});
-(FateAccelerated as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.FateAccelerated
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const FateOfCthulhu = Template.bind({});
-(FateOfCthulhu as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.FateOfCthulhu
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const DresdenFilesAccelerated = Template.bind({});
-(DresdenFilesAccelerated as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.DresdenFilesAccelerated
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const VentureCity = Template.bind({});
-(VentureCity as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.VentureCity
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const Heartbreaker = Template.bind({});
-(Heartbreaker as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.Heartbreaker
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const IronEddaAccelerated = Template.bind({});
-(IronEddaAccelerated as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.IronEddaAccelerated
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const Maze = Template.bind({});
-(Maze as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(CharacterTemplates.Maze);
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const Dnd5e = Template.bind({});
-(Dnd5e as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(CharacterTemplates.Dnd5e);
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const TheWitchIsDead = Template.bind({});
-(TheWitchIsDead as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.TheWitchIsDead
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const EdgeOfTheEmpire = Template.bind({});
-(EdgeOfTheEmpire as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.EdgeOfTheEmpire
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const EdgeOfTheEmpire_FR = Template.bind({});
-(EdgeOfTheEmpire_FR as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.EdgeOfTheEmpire_FR
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const EvolutionPulseHydrah = Template.bind({});
-(EvolutionPulseHydrah as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.EvolutionPulse_Hydrah
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const EvolutionPulse_Hyonos = Template.bind({});
-(EvolutionPulse_Hyonos as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.EvolutionPulse_Hyonos
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const EvolutionPulse_LostH = Template.bind({});
-(EvolutionPulse_LostH as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.EvolutionPulse_LostH
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const EvolutionPulse_Obscura = Template.bind({});
-(EvolutionPulse_Obscura as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.EvolutionPulse_Obscura
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const EvolutionPulse_Proxy = Template.bind({});
-(EvolutionPulse_Proxy as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.EvolutionPulse_Proxy
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const ThePool = Template.bind({});
-(ThePool as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(CharacterTemplates.ThePool);
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const TunnelsAndTrolls = Template.bind({});
-(TunnelsAndTrolls as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.TunnelsAndTrolls
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const StrandsOfFate = Template.bind({});
-(StrandsOfFate as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(
-      CharacterTemplates.StrandsOfFate
-    );
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
-export const Blank = Template.bind({});
-(Blank as any).loaders = [
-  async () => {
-    const character = await CharacterFactory.make(CharacterTemplates.Blank);
-    return { character: overrideCharacterDateForStorybook(character) };
-  },
-];
+// prettier-ignore
+export const FateCondensed = makeCharacterSheetStory(CharacterTemplates.FateCondensed)
+// prettier-ignore
+export const FateCore = makeCharacterSheetStory(CharacterTemplates.FateCore)
+// prettier-ignore
+export const FateAccelerated = makeCharacterSheetStory(CharacterTemplates.FateAccelerated)
+// prettier-ignore
+export const FateOfCthulhu = makeCharacterSheetStory(CharacterTemplates.FateOfCthulhu)
+// prettier-ignore
+export const DresdenFilesAccelerated = makeCharacterSheetStory(CharacterTemplates.DresdenFilesAccelerated)
+// prettier-ignore
+export const VentureCity = makeCharacterSheetStory(CharacterTemplates.VentureCity)
+// prettier-ignore
+export const Heartbreaker = makeCharacterSheetStory(CharacterTemplates.Heartbreaker)
+// prettier-ignore
+export const IronEddaAccelerated = makeCharacterSheetStory(CharacterTemplates.IronEddaAccelerated)
+// prettier-ignore
+export const StrandsOfFate = makeCharacterSheetStory(CharacterTemplates.StrandsOfFate)
+// prettier-ignore
+export const EvolutionPulse_Hydrah = makeCharacterSheetStory(CharacterTemplates.EvolutionPulse_Hydrah)
+// prettier-ignore
+export const EvolutionPulse_Hyonos = makeCharacterSheetStory(CharacterTemplates.EvolutionPulse_Hyonos)
+// prettier-ignore
+export const EvolutionPulse_LostH = makeCharacterSheetStory(CharacterTemplates.EvolutionPulse_LostH)
+// prettier-ignore
+export const EvolutionPulse_Obscura = makeCharacterSheetStory(CharacterTemplates.EvolutionPulse_Obscura)
+// prettier-ignore
+export const EvolutionPulse_Proxy = makeCharacterSheetStory(CharacterTemplates.EvolutionPulse_Proxy)
+// prettier-ignore
+export const FateCoreSpanish = makeCharacterSheetStory(CharacterTemplates.FateCoreSpanish)
+// prettier-ignore
+export const FateCondensedSpanish = makeCharacterSheetStory(CharacterTemplates.FateCondensedSpanish)
+// prettier-ignore
+export const MärchenkriegerLOS = makeCharacterSheetStory(CharacterTemplates.MärchenkriegerLOS)
+// prettier-ignore
+export const TachyonSquadronCharacter = makeCharacterSheetStory(CharacterTemplates.TachyonSquadronCharacter)
+// prettier-ignore
+export const TachyonSquadronShip = makeCharacterSheetStory(CharacterTemplates.TachyonSquadronShip)
+// prettier-ignore
+export const TachyonSquadronCharacterAndShip = makeCharacterSheetStory(CharacterTemplates.TachyonSquadronCharacterAndShip)
+// prettier-ignore
+export const DresdenFilesRPGCharacter = makeCharacterSheetStory(CharacterTemplates.DresdenFilesRPGCharacter)
+// prettier-ignore
+export const DresdenFilesRPGSpellCaster = makeCharacterSheetStory(CharacterTemplates.DresdenFilesRPGSpellCaster)
+// prettier-ignore
+export const DresdenFilesRPGVampire = makeCharacterSheetStory(CharacterTemplates.DresdenFilesRPGVampire)
+// prettier-ignore
+export const ThePool = makeCharacterSheetStory(CharacterTemplates.ThePool)
+// prettier-ignore
+export const TunnelsAndTrolls = makeCharacterSheetStory(CharacterTemplates.TunnelsAndTrolls)
+// prettier-ignore
+export const Maze = makeCharacterSheetStory(CharacterTemplates.Maze)
+// prettier-ignore
+export const Dnd5e = makeCharacterSheetStory(CharacterTemplates.Dnd5e)
+// prettier-ignore
+export const TheWitchIsDead = makeCharacterSheetStory(CharacterTemplates.TheWitchIsDead)
+// prettier-ignore
+export const EdgeOfTheEmpire = makeCharacterSheetStory(CharacterTemplates.EdgeOfTheEmpire)
+// prettier-ignore
+export const EdgeOfTheEmpire_FR = makeCharacterSheetStory(CharacterTemplates.EdgeOfTheEmpire_FR)
+// prettier-ignore
+export const Blank = makeCharacterSheetStory(CharacterTemplates.Blank)
+
 export const LoremIpsum = Template.bind({});
 LoremIpsum.args = {
   character: CharacterFactory.migrate(LoremIpsumTemplate as any),
 };
 
-function overrideCharacterDateForStorybook(character: ICharacter): ICharacter {
-  return {
-    ...character,
-    id: "50fa2",
-    lastUpdated: dayJS("2021-01-01").unix(),
-  };
+function makeCharacterSheetStory(template: CharacterTemplates) {
+  const story = Template.bind({});
+  (story as any).loaders = [
+    async () => {
+      const character = await CharacterFactory.make(template);
+      return { character: overrideCharacterDateForStorybook(character) };
+    },
+  ];
+  return story;
+
+  function overrideCharacterDateForStorybook(
+    character: ICharacter
+  ): ICharacter {
+    return {
+      ...character,
+      id: "50fa2",
+      lastUpdated: dayJS("2021-01-01").unix(),
+    };
+  }
 }